save current CPU state into json file
Neither of the 'native' nim solutions (including an "official" one (this
more like leftovers from Aporia editor), in
~nim-lang/dialogs~ [fn:dialogs-4]) were working (as in - first is
unusable, second does not compile), so I decided to use the first
library from the =stackoverflow= recommendations. Surprisingly enough,
it worked and was easy to use. I wouldn't be so sure about windows
support, but on linux it does the job just fine.

[fn:dialogs-4] It implies I use GTK2 for the application, because it's
API excepts me to have ~PWindow~ available